Move stats functions to stats.rs
